backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
google-genai==0.3.0
python-dotenv==1.0.1
slowapi==0.1.9
